ml-service/requirements.txt
fastapi
uvicorn
pandas
scikit-learn
joblib
groq
pydantic
numpy
python-dotenv  